# api-docs/openapi.yaml
openapi: 3.0.3
info:
  title: 供应链金融系统 API文档
  description: 基于FISCO BCOS区块链的供应链金融系统接口文档
  contact:
    name: 开发团队
  version: 1.0.0
servers:
- url: http://localhost:8080
  description: Inferred Url
...
paths:
...
  /api/audit/user/{userAddress}:
    get:
      tags:
      - 审计日志管理
      summary: 查询用户的操作历史
      description: 查询特定用户的所有操作记录
      operationId: getUserHistoryUsingGET
      parameters:
      - name: userAddress
        in: path
        description: 用户地址
        required: true
        style: simple
        schema:
          type: string
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/Result«List«审计日志»»'
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Not Found
      security:
      - Authorization:
        - global
...
  /api/relations/user/{userId}:
    get:
      tags:
      - 企业和用户关联查询
      summary: 获取用户及其企业信息
      description: 根据用户ID查询用户详情和所属企业信息
      operationId: getUserWithEnterpriseUsingGET
      parameters:
      - name: userId
        in: path
        description: 用户ID
        required: true
        style: simple
        schema:
          type: integer
          format: int64
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/Result«用户企业信息»'
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Not Found
      security:
      - Authorization:
        - global
...
  /api/users/{userId}:
    get:
      tags:
      - 用户管理
      summary: 获取用户信息
      description: 根据ID查询用户详细信息
      operationId: getUserUsingGET
      parameters:
      - name: userId
        in: path
        description: 用户ID
        required: true
        style: simple
        schema:
          type: integer
          format: int64
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/Result«User»'
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Not Found
      security:
      - Authorization:
        - global
    put:
      tags:
      - 用户管理
      summary: 更新用户信息
      description: 更新指定用户的信息
      operationId: updateUserUsingPUT
      parameters:
      - name: authenticated
        in: query
        required: false
        style: form
        schema:
          type: boolean
      - name: authorities[0].authority
        in: query
        required: false
        style: form
        schema:
          type: string
      - name: credentials
        in: query
        required: false
        style: form
        schema:
          type: object
      - name: details
        in: query
        required: false
        style: form
        schema:
          type: object
      - name: principal
        in: query
        required: false
        style: form
        schema:
          type: object
      - name: userId
        in: path
        description: 用户ID
        required: true
        style: simple
        schema:
          type: integer
          format: int64
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateUserRequest'
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/Result«User»'
        '201':
          description: Created
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Not Found
      security:
      - Authorization:
        - global
    delete:
      tags:
      - 用户管理
      summary: 删除用户
      description: 删除指定用户
      operationId: deleteUserUsingDELETE
      parameters:
      - name: authenticated
        in: query
        required: false
        style: form
        schema:
          type: boolean
      - name: authorities[0].authority
        in: query
        required: false
        style: form
        schema:
          type: string
      - name: credentials
        in: query
        required: false
        style: form
        schema:
          type: object
      - name: details
        in: query
        required: false
        style: form
        schema:
          type: object
      - name: principal
        in: query
        required: false
        style: form
        schema:
          type: object
      - name: userId
        in: path
        description: 用户ID
        required: true
        style: simple
        schema:
          type: integer
          format: int64
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/Result«string»'
        '204':
          description: No Content
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
      security:
      - Authorization:
        - global
...
  /api/warehouse-receipt:
    post:
      tags:
      - 仓单管理
      summary: 创建仓单
      description: 货主创建仓单，使用CreateReceiptParams结构体封装参数，避免Solidity 16变量限制
      operationId: createReceiptUsingPOST
      parameters:
      - name: authenticated
        in: query
        required: false
        style: form
        schema:
          type: boolean
      - name: authorities[0].authority
        in: query
        required: false
        style: form
        schema:
          type: string
      - name: credentials
        in: query
        required: false
        style: form
        schema:
          type: object
      - name: details
        in: query
        required: false
        style: form
        schema:
          type: object
      - name: principal
        in: query
        required: false
        style: form
        schema:
          type: object
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/创建仓单请求'
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/Result«仓单»'
        '201':
          description: Created
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Not Found
      security:
      - Authorization:
        - global
  /api/warehouse-receipt/{receiptId}:
    get:
      tags:
      - 仓单管理
      summary: 获取仓单详情
      description: 根据仓单ID查询详细信息
      operationId: getReceiptUsingGET
      parameters:
      - name: receiptId
        in: path
        description: 仓单ID
        required: true
        style: simple
        schema:
          type: string
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/Result«仓单»'
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Not Found
      security:
      - Authorization:
        - global
  /api/warehouse-receipt/{receiptId}/release:
    post:
      tags:
      - 仓单管理
      summary: 释放仓单
      description: 释放已质押的仓单
      operationId: releaseReceiptUsingPOST
      parameters:
      - name: receiptId
        in: path
        description: 仓单ID
        required: true
        style: simple
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/仓单释放请求'
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/Result«仓单释放响应»'
        '201':
          description: Created
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Not Found
      security:
      - Authorization:
        - global
  /api/warehouse-receipt/{receiptId}/releases:
    get:
      tags:
      - 仓单管理
      summary: 获取仓单释放历史
      description: 查询仓单的所有释放记录
      operationId: getReleaseHistoryUsingGET
      parameters:
      - name: receiptId
        in: path
        description: 仓单ID
        required: true
        style: simple
        schema:
          type: string
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/Result«List«仓单释放记录»»'
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Not Found
      security:
      - Authorization:
        - global
components:
  schemas:
...
    Result«List«仓单释放记录»»:
      title: Result«List«仓单释放记录»»
      type: object
      properties:
        code:
          type: integer
          format: int32
        data:
          type: array
          items:
            $ref: '#/components/schemas/仓单释放记录'
        message:
          type: string
        timestamp:
          type: integer
          format: int64
...
    Result«List«审计日志»»:
      title: Result«List«审计日志»»
      type: object
      properties:
        code:
          type: integer
          format: int32
        data:
          type: array
          items:
            $ref: '#/components/schemas/审计日志'
        message:
          type: string
        timestamp:
          type: integer
          format: int64
...
    Result«User»:
      title: Result«User»
      type: object
      properties:
        code:
          type: integer
          format: int32
        data:
          $ref: '#/components/schemas/User'
        message:
          type: string
        timestamp:
          type: integer
          format: int64
...
    Result«string»:
      title: Result«string»
      type: object
      properties:
        code:
          type: integer
          format: int32
        data:
          type: string
        message:
          type: string
        timestamp:
          type: integer
          format: int64
    Result«仓单»:
      title: Result«仓单»
      type: object
      properties:
        code:
          type: integer
          format: int32
        data:
          $ref: '#/components/schemas/仓单'
        message:
          type: string
        timestamp:
          type: integer
          format: int64
    Result«仓单释放响应»:
      title: Result«仓单释放响应»
      type: object
      properties:
        code:
          type: integer
          format: int32
        data:
          $ref: '#/components/schemas/仓单释放响应'
        message:
          type: string
        timestamp:
          type: integer
          format: int64
...
    Result«用户企业信息»:
      title: Result«用户企业信息»
      type: object
      properties:
        code:
          type: integer
          format: int32
        data:
          $ref: '#/components/schemas/用户企业信息'
        message:
          type: string
        timestamp:
          type: integer
          format: int64
...
    UpdateUserRequest:
      title: UpdateUserRequest
      required:
      - username
      type: object
      properties:
        avatarUrl:
          type: string
          description: 头像URL
        department:
          type: string
          description: 部门
        email:
          type: string
          description: 电子邮箱
        phone:
          type: string
          description: 手机号码
        position:
          type: string
          description: 职位
        realName:
          type: string
          description: 真实姓名
        username:
          type: string
          description: 用户名
    User:
      title: User
      required:
      - status
      - userType
      - username
      type: object
      properties:
        active:
          type: boolean
        avatarUrl:
          type: string
          description: 头像URL
        department:
          type: string
          description: 部门
          example: 财务部
        disabled:
          type: boolean
        email:
          type: string
          description: 电子邮箱
          example: zhangsan@example.com
        enterpriseId:
          type: integer
          description: 所属企业ID
          format: int64
          example: 1
        invitationCode:
          type: string
          description: 注册时使用的邀请码
          example: INV2024011812345678
        locked:
          type: boolean
        phone:
          type: string
          description: 手机号码
          example: '13800138000'
        position:
          type: string
          description: 职位
          example: 财务经理
        realName:
          type: string
          description: 真实姓名
          example: 张三
        registrationRemarks:
          type: string
          description: 注册备注信息
          example: 用户注册申请备注
        status:
          type: string
          description: 用户状态
          enum:
          - ACTIVE
          - DISABLED
          - LOCKED
          - PENDING
        userType:
          type: string
          description: 用户类型
          enum:
          - ADMIN
          - AUDITOR
          - ENTERPRISE_USER
          - OPERATOR
        username:
          type: string
          description: 用户名（登录账号）
          example: zhangsan
...
    仓单:
      title: 仓单
      required:
      - createdAt
      - expiryDate
      - goodsName
      - ownerAddress
      - quantity
      - receiptId
      - status
      - storageDate
      - totalPrice
      - unit
      - unitPrice
      - warehouseAddress
      type: object
      properties:
        createdAt:
          type: string
          description: 创建时间
          format: date-time
        expiryDate:
          type: string
          description: 到期日期
          format: date-time
        financeAmount:
          type: number
          description: 融资金额
          format: bigdecimal
          example: 350000.0
        financeDate:
          type: string
          description: 融资日期
          format: date-time
        financeRate:
          type: integer
          description: 融资利率（基点）
          format: int32
          example: 500
        financialInstitution:
          type: string
          description: 金融机构地址
          example: '0x9876543210987654321098765432109876543210'
        goodsName:
          type: string
          description: 货物名称
          example: 钢材
        goodsType:
          type: string
          description: 货物类型
          example: 螺纹钢
        origin:
          type: string
          description: 产地
          example: 上海
        ownerAddress:
          type: string
          description: 所有者地址
          example: '0x1234567890abcdef1234567890abcdef12345678'
        pledgeAmount:
          type: number
          description: 质押金额
          format: bigdecimal
          example: 400000.0
        quality:
          type: string
          description: 质量等级
          example: 一级
        quantity:
          type: number
          description: 数量
          format: bigdecimal
          example: 100.0
        receiptId:
          type: string
          description: 仓单ID
          example: receipt001
        releaseDate:
          type: string
          description: 释放日期
          format: date-time
        status:
          type: string
          description: 仓单状态
          example: CREATED
          enum:
          - CREATED
          - EXPIRED
          - FINANCED
          - LIQUIDATED
          - PLEDGED
          - RELEASED
          - VERIFIED
        storageDate:
          type: string
          description: 入库日期
          format: date-time
        totalPrice:
          type: number
          description: 总价
          format: bigdecimal
          example: 500000.0
        txHash:
          type: string
          description: 区块链交易哈希
          example: '0xabcdef1234567890abcdef1234567890abcdef1234567890abcdef1234567890'
        unit:
          type: string
          description: 单位
          example: 吨
        unitPrice:
          type: number
          description: 单价
          format: bigdecimal
          example: 5000.0
        updatedAt:
          type: string
          description: 更新时间
          format: date-time
        warehouseAddress:
          type: string
          description: 仓库地址
          example: '0xabcdefabcdefabcdefabcdefabcdefabcdefabcd'
        warehouseLocation:
          type: string
          description: 仓库位置
          example: 上海市宝山区仓库A区01号货架
      description: 仓单信息实体
    仓单释放响应:
      title: 仓单释放响应
      type: object
      properties:
        financeAmount:
          type: number
          description: 融资金额
          format: bigdecimal
        financeDate:
          type: string
          description: 融资日期
          format: date-time
        financeRate:
          type: integer
          description: 融资利率
          format: int32
        financialInstitutionAddress:
          type: string
          description: 金融机构地址
        id:
          type: integer
          description: 释放记录ID
          format: int64
        interestAmount:
          type: number
          description: 利息金额
          format: bigdecimal
        ownerAddress:
          type: string
          description: 所有者地址
        pledgeAmount:
          type: number
          description: 质押金额
          format: bigdecimal
        receiptId:
          type: string
          description: 仓单ID
        releaseDate:
          type: string
          description: 释放日期
          format: date-time
        releaseType:
          type: string
          description: 释放类型
          enum:
          - FULL_REPAYMENT
          - LIQUIDATION
          - MANUAL
          - MATURITY
          - PARTIAL_REPAYMENT
        remark:
          type: string
          description: 释放备注
        repaymentAmount:
          type: number
          description: 还款金额
          format: bigdecimal
        txHash:
          type: string
          description: 区块链交易哈希
      description: 仓单释放操作的结果
    仓单释放记录:
      title: 仓单释放记录
      required:
      - createdAt
      - ownerAddress
      - receiptId
      - releaseDate
      - releaseType
      type: object
      properties:
        createdAt:
          type: string
          description: 创建时间
          format: date-time
        financeAmount:
          type: number
          description: 融资金额
          format: bigdecimal
          example: 350000.0
        financeDate:
          type: string
          description: 融资日期
          format: date-time
        financeRate:
          type: integer
          description: 融资利率（基点）
          format: int32
          example: 500
        financialInstitutionAddress:
          type: string
          description: 金融机构地址
          example: '0x9876543210987654321098765432109876543210'
        interestAmount:
          type: number
          description: 利息金额
          format: bigdecimal
          example: 5000.0
        ownerAddress:
          type: string
          description: 仓单所有者地址
          example: '0x1234567890abcdef1234567890abcdef12345678'
        pledgeAmount:
          type: number
          description: 质押金额
          format: bigdecimal
          example: 400000.0
        receiptId:
          type: string
          description: 仓单ID
          example: receipt001
        releaseDate:
          type: string
          description: 释放日期
          format: date-time
        releaseType:
          type: string
          description: 释放类型
          example: FULL_REPAYMENT
          enum:
          - FULL_REPAYMENT
          - LIQUIDATION
          - MANUAL
          - MATURITY
          - PARTIAL_REPAYMENT
        remark:
          type: string
          description: 释放备注
          example: 全额还款后释放仓单
        repaymentAmount:
          type: number
          description: 还款金额
          format: bigdecimal
          example: 355000.0
        txHash:
          type: string
          description: 区块链交易哈希
          example: '0xabcdef1234567890abcdef1234567890abcdef1234567890abcdef1234567890'
      description: 仓单质押释放操作记录实体
    仓单释放请求:
      title: 仓单释放请求
      required:
      - releaseType
      type: object
      properties:
        interestAmount:
          type: number
          description: 利息金额
          format: bigdecimal
          example: 50000.0
        releaseType:
          type: string
          description: 释放类型
          example: FULL_REPAYMENT
          enum:
          - FULL_REPAYMENT
          - LIQUIDATION
          - MANUAL
          - MATURITY
          - PARTIAL_REPAYMENT
        remark:
          type: string
          description: 释放备注
          example: 全额还款，解除质押
        repaymentAmount:
          type: number
          description: 还款金额（全额还款时可选）
          format: bigdecimal
          example: 1000000.0
      description: 用于仓单质押释放的请求参数
...
    创建仓单请求:
      title: 创建仓单请求
      required:
      - expiryDate
      - goods
      - receiptId
      - storageDate
      - warehouseAddress
      - warehouseLocation
      type: object
      properties:
        expiryDate:
          type: string
          description: 过期日期
          format: date-time
        goods:
          description: 货物信息
          $ref: '#/components/schemas/货物信息'
        receiptId:
          type: string
          description: 仓单ID
          example: WR20240113001
        storageDate:
          type: string
          description: 入库日期
          format: date-time
        warehouseAddress:
          type: string
          description: 仓库地址
          example: '0x567890abcdef1234'
        warehouseLocation:
          type: string
          description: 仓库物理位置
          example: 上海市浦东新区XX仓库
      description: 用于创建仓单的请求参数
...
    审计日志:
      title: 审计日志
      required:
      - actionDesc
      - actionType
      - createdAt
      - isSuccess
      - module
      type: object
      properties:
        actionDesc:
          type: string
          description: 操作描述
          example: 创建票据
        actionType:
          type: string
          description: 操作类型
          example: CREATE
        changedFields:
          type: string
          description: 变更的字段
        createdAt:
          type: string
          description: 创建时间
          format: date-time
        duration:
          type: integer
          description: 操作时长（毫秒）
          format: int64
          example: 150
        entityId:
          type: string
          description: 实体ID
          example: bill001
        entityType:
          type: string
          description: 实体类型
          example: Bill
        errorMessage:
          type: string
          description: 错误信息
        isSuccess:
          type: boolean
          description: 操作是否成功
          example: true
        module:
          type: string
          description: 操作模块
          example: BILL
        newValue:
          type: string
          description: 操作后的数据
        oldValue:
          type: string
          description: 操作前的数据
        requestIp:
          type: string
          description: 请求IP地址
          example: 192.168.1.100
        requestMethod:
          type: string
          description: 请求方法
          example: POST
        requestUrl:
          type: string
          description: 请求URL
          example: /api/bill/create
        result:
          type: string
          description: 操作结果
          example: SUCCESS
        tenantId:
          type: string
          description: 租户ID
          example: tenant001
        txHash:
          type: string
          description: 区块链交易哈希
          example: '0xabcdef1234567890abcdef1234567890abcdef1234567890abcdef1234567890'
        userAddress:
          type: string
          description: 操作人地址
          example: '0x1234567890abcdef1234567890abcdef12345678'
        userAgent:
          type: string
          description: 用户代理
          example: Mozilla/5.0...
        userName:
          type: string
          description: 操作人姓名
          example: 张三
      description: 系统操作审计日志实体
...
    用户企业信息:
      title: 用户企业信息
      type: object
      properties:
        creditRating:
          type: integer
          description: 信用评级
          format: int32
        department:
          type: string
          description: 部门
        email:
          type: string
          description: 邮箱
        enterpriseAddress:
          type: string
          description: 企业地址
        enterpriseId:
          type: integer
          description: 企业ID
          format: int64
        enterpriseName:
          type: string
          description: 企业名称
        enterpriseRole:
          type: string
          description: 企业角色
        enterpriseStatus:
          type: string
          description: 企业状态
        lastLoginTime:
          type: string
          description: 最后登录时间
        loginCount:
          type: integer
          description: 登录次数
          format: int32
        phone:
          type: string
          description: 手机号
        position:
          type: string
          description: 职位
        realName:
          type: string
          description: 真实姓名
        status:
          type: string
          description: 用户状态
        userId:
          type: integer
          description: 用户ID
          format: int64
        userType:
          type: string
          description: 用户类型
        username:
          type: string
          description: 用户名
      description: 用户及其所属企业的详细信息
...
    货物信息:
      title: 货物信息
      required:
      - goodsName
      - quantity
      - totalPrice
      - unit
      - unitPrice
      type: object
      properties:
        goodsName:
          type: string
          description: 货物名称
          example: 钢材
        goodsType:
          type: string
          description: 货物类型
          example: 建材
        origin:
          type: string
          description: 产地
          example: 上海
        quality:
          type: string
          description: 质量等级
          example: 一级
        quantity:
          type: number
          description: 数量
          format: bigdecimal
          example: 100.0
        totalPrice:
          type: number
          description: 总价（分）
          format: bigdecimal
          example: 500000.0
        unit:
          type: string
          description: 单位
          example: 吨
        unitPrice:
          type: number
          description: 单价（分）
          format: bigdecimal
          example: 5000.0
      description: 仓单货物的详细信息
...
  securitySchemes:
    Authorization:
      type: apiKey
      name: Authorization
      in: header